+++ crates/core/executor/src/events/precompiles/sha256_compress.rs
@@ -11,7 +11,7 @@ use crate::events::{
 #[derive(Default, Debug, Clone, Serialize, Deserialize)]
 pub struct ShaCompressEvent {
     /// The shard number.
-    pub shard: u32,
+    pub shard: u16,
     /// The clock cycle.
     pub clk: u32,
     /// The pointer to the word.
